+++ src/noteoverview.ts
@@ -18,6 +18,7 @@ let globalSettings: any = {};
 const consoleLogLevel = "verbose";
 let firstSyncCompleted = false;
 let joplinNotebooks: any = null;
+let logFile = null;
 
 export namespace noteoverview {
   export async function getImageNr(
@@ -25,7 +26,7 @@ export namespace noteoverview {
     imagrNr: number,
     imageSettings: Object
   ): Promise<string> {
-    logging.silly("func: getImageNr");
+    logging.verbose("func: getImageNr");
     const regExresourceId = /!\[([^\]]+|)\]\(:\/(?<resourceId>[\da-z]{32})\)/g;
     let ids = [];
     let imageId = null;
@@ -106,6 +107,13 @@ export namespace noteoverview {
     noteoverviewSettings: object
   ): Promise<string> {
     let settingsBlock = [];
+
+    // Replace search with original search
+    if (noteoverviewSettings["searchWithVars"]) {
+      noteoverviewSettings["search"] = noteoverviewSettings["searchWithVars"];
+      delete noteoverviewSettings["searchWithVars"];
+    }
+
     const yamlBlock = YAML.stringify(noteoverviewSettings);
     settingsBlock.push("<!-- note-overview-plugin");
     settingsBlock.push(yamlBlock.substring(0, yamlBlock.length - 1));
@@ -183,7 +191,7 @@ export namespace noteoverview {
     todo_completed: number,
     type: string
   ): Promise<string> {
-    logging.silly("func: getToDoDateColor");
+    logging.verbose("func: getToDoDateColor");
     const now = new Date();
     let colorType = "";
 
@@ -291,7 +299,7 @@ export namespace noteoverview {
     todo_due: number,
     todo_completed: number
   ) {
-    logging.silly("func: getToDoStatus");
+    logging.verbose("func: getToDoStatus");
     const now = new Date();
     if (todo_completed === 0 && todo_due !== 0 && todo_due < now.getTime())
       return "overdue";
@@ -453,9 +461,9 @@ export namespace noteoverview {
   }
 
   export async function loadNotebooks(reload = false) {
-    logging.silly("Func: loadNotebooks");
+    logging.verbose("Func: loadNotebooks");
     if (reload === true || joplinNotebooks === null) {
-      logging.silly("load notebooks");
+      logging.verbose("load notebooks");
       joplinNotebooks = {};
       let queryFolders;
       let pageQuery = 1;
@@ -702,7 +710,15 @@ export namespace noteoverview {
         return;
       }
 
+      noteOverviewSettings["searchWithVars"] = noteOverviewSettings["search"];
+      noteOverviewSettings["search"] = await noteoverview.replaceSearchVars(
+        noteOverviewSettings["search"]
+      );
+
       logging.verbose("Search: " + noteOverviewSettings["search"]);
+      logging.verbose(
+        "Search with vars: " + noteOverviewSettings["searchWithVars"]
+      );
 
       // add original content before the settings block
       if (startOrgTextIndex != startIndex) {
@@ -752,7 +768,7 @@ export namespace noteoverview {
   export async function getOptions(
     overviewSettings: any
   ): Promise<OverviewOptions> {
-    logging.silly("func: getOptions");
+    logging.verbose("func: getOptions");
     const settings: any = {};
     settings.overview = overviewSettings;
 
@@ -845,7 +861,7 @@ export namespace noteoverview {
     noteTitle: string,
     overviewSettings: any
   ): Promise<string[]> {
-    logging.silly("func: getOverviewContent");
+    logging.verbose("func: getOverviewContent");
     const query: string = overviewSettings["search"];
     let overviewContent: string[] = [];
 
@@ -1100,12 +1116,33 @@ export namespace noteoverview {
     return "|" + info.join("|") + "|";
   }
 
+  export async function removeNoteoverviewCode(data: string): Promise<string> {
+    data = data.replace(
+      /(?<!```\n)(?<!``` \n)(<!--\s?note-overview-plugin([\w\W]*?)-->)/gi,
+      "REMOVE_NOTOVERVIEW_LINE"
+    );
+    data = data.replace(
+      /(<!--endoverview-->)(?!\n```)/gi,
+      "REMOVE_NOTOVERVIEW_LINE"
+    );
+
+    const lines = data.split("\n");
+    let newLines = [];
+    for (const line of lines) {
+      if (line.match("REMOVE_NOTOVERVIEW_LINE") === null) {
+        newLines.push(line);
+      }
+    }
+
+    return newLines.join("\n");
+  }
+
   export async function getFieldValue(
     field: string,
     fields: any,
     options: OverviewOptions
   ): Promise<string> {
-    logging.silly("func: getFieldValue for " + field);
+    logging.verbose("func: getFieldValue for " + field);
     let value = "";
     switch (field) {
       case "title":
@@ -1203,6 +1240,8 @@ export namespace noteoverview {
         value = fields[field];
     }
 
+    value = await noteoverview.removeNoteoverviewCode(value);
+
     if (options.escapeForTable === true) {
       value = await noteoverview.escapeForTable(value);
     }
@@ -1234,26 +1273,15 @@ export namespace noteoverview {
     return await noteoverview.removeNewLineAt(orgContent, stripe[0], stripe[1]);
   }
 
-  export async function getFileLogLevel(): Promise<any> {
-    const logLevelFile = path.join(
-      await joplin.plugins.installationDir(),
-      "debug.txt"
-    );
-    if (fs.existsSync(logLevelFile)) {
-      return "silly";
-    } else {
-      return "error";
-    }
-  }
-
   export async function setupLogging() {
     const logFormatFile = "[{y}-{m}-{d} {h}:{i}:{s}.{ms}] [{level}] {text}";
     const logFormatConsole = "[{level}] {text}";
-    const logFile = path.join(
+    logFile = path.join(
       await joplin.plugins.installationDir(),
       "noteoverview.log"
     );
-    const levelFile = await noteoverview.getFileLogLevel();
+
+    const levelFile = await joplin.settings.value("fileLogLevel");
     logging.transports.file.format = logFormatFile;
     logging.transports.file.level = levelFile;
     logging.transports.file.resolvePath = () => logFile;
@@ -1261,10 +1289,22 @@ export namespace noteoverview {
     logging.transports.console.format = logFormatConsole;
   }
 
+  export async function deleteLogFile() {
+    logging.verbose("Delete log file");
+    if (fs.existsSync(logFile)) {
+      try {
+        await fs.unlinkSync(logFile);
+      } catch (e) {
+        logging.error("deleteLogFile: " + e.message);
+      }
+    }
+  }
+
   export async function init() {
     logging.info("Note overview plugin started!");
 
     await settings.register();
+    await noteoverview.deleteLogFile();
     await noteoverview.setupLogging();
 
     noteoverviewDialog = await joplin.views.dialogs.create(
@@ -1331,6 +1371,10 @@ export namespace noteoverview {
         await joplin.settings.value("updateInterval")
       );
     }
+
+    if (event.keys.indexOf("fileLogLevel") !== -1) {
+      await noteoverview.setupLogging();
+    }
   }
 
   export async function setTimer(updateInterval: number) {
@@ -1353,6 +1397,51 @@ export namespace noteoverview {
     } else {
       timer = null;
     }
+  }
+
+  export async function replaceSearchVars(query: string): Promise<string> {
+    logging.verbose("replaceSearchVars");
+
+    const joplinLocale = await joplin.settings.globalValue("locale");
+    const momentsLocale = joplinLocale.split("_")[0];
+
+    return query.replace(/{{moments:(?<format>[^}]+)}}/g, (match, groups) => {
+      let now = new Date(Date.now());
+      let momentDate = moment(now);
+      momentDate.locale(momentsLocale);
+
+      // Modify date
+      const modifyDateRegEx = /( modify:)(?<modify>.*)/;
+      const modifyDate = groups.match(modifyDateRegEx);
+      groups = groups.replace(modifyDateRegEx, "");
+      if (modifyDate !== null) {
+        let actions = [];
+        if (modifyDate["groups"]["modify"].match(",") !== null) {
+          actions = modifyDate["groups"]["modify"].split(",");
+        } else {
+          actions.push(modifyDate["groups"]["modify"]);
+        }
+
+        for (const action of actions) {
+          let add = action.substring(0, 1);
+          let quantity = action.substring(1, action.length - 1);
+          let type = action.substring(action.length - 1, action.length);
+
+          try {
+            if (add == "-") {
+              momentDate.subtract(quantity, type);
+            } else if (add == "+") {
+              momentDate.add(quantity, type);
+            }
+          } catch (e) {
+            logging.error(e);
+          }
+        }
+        now = new Date(momentDate.valueOf());
+      }
+
+      return momentDate.format(groups);
+    });
   }
 }